Deal with int overflow in matrixProcs
This clamps the value to be packed to 2^14-1 if the value + 1
would exceed 2^14-1.

This adds a macro for unlikely branches as an attempt to
reduce the performance hit of this extra check.

I was unable to reproduce the overflow from the linked bug
higher upstack, but this should squash it.

Change-Id: I47a9fc6faa5fd22763c424b4e6e4041f569dcb74
Bug: 1357122
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/577197
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>