Remove implicit prototypes from DSLFunction.
For convenience, DSLFunction was designed to automatically insert a
prototype into the program as soon as it was instantiated; this
prototype was subsequently removed if you immediately declared a
function.

This convenience behavior made sense if you assume a world where we
commonly use DSL to implement all programs. But in practice, outside
of unit tests, DSLFunction is _only_ used in the Parser, where this
feature doesn't actually improve convenience; it just wastes IRNodes.

Now, we can use `func.prototype();` to insert a prototype into the
program.

Change-Id: I1e297715758752cd4eebaf3a0bded7dfdb6138d2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/578816
Commit-Queue: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>