[graphite] Add DenseBiMap template to consolidate index tracking in DrawPass
The TextureBindings, the uniforms for paint and renderstep, the final
proxies and sampler descriptions, and the graphics pipeline descriptions
all were being mapped from their type to a dense index that was then
used to handle state transitions and bind commands.

This moves all of that logic into a template.

Change-Id: Icf2c3a2b013295751071ace8275c75606b56014a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/576577
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>