Convert drawPaint calls to clears only when it's safe.
If the paint has a shader, is non-opaque, has a weird blend mode set,
etc., it cannot be represented as a clear.

If the paint has a color filter, it _can_ be represented as a clear,
but we need to pass the paint color through the filter first.

Change-Id: I618c08df425603d9cee5bc039ce2582ddaa438e5
Bug: skia:13731
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/578856
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>