In legacy 8888 sprite blitters, use saturating adds
This fixes blending errors if the incoming bitmap has out-of-gamut
colors (RGB > A), where the color channels were wrapping to zero.
Not sure about the NEON perf impact. For all of the various Intel
instruction sets, this should only add one cycle, if my read of the
docs is correct.

No tests today encounter this, but upcoming changes relax clamping
rules in raster pipeline, and the readpixels GM produced incorrect
results (because it generates a premul bitmap with out-of-gamut
colors). There's nothing stopping clients from hitting this bug
today, other than luck, and the fact that the sprite blitters only
handle very specific circumstances.

This also fixes numerous egregious problems in narrow-8888, as well
as the linked Chromium bug, where bookmark/tab icons with saturated
colors would end up out-of-gamut after color management.

The change to SkPMSrcOver is going to cause a performance regression,
but *hopefully* it's not severe. The bulk of sprite draws should be
spent in the SIMD fast-paths, so the scalar code is largely used for
tail pixels.

Example error fixed by this CL:
  https://gold.skia.org/detail?test=readpixels&digest=e2f109c4b2a27477a64e1cffc22e0fba

Bug: skia:13715
Bug: chromium:1406521
Change-Id: I4029705dd5ef1fe077995c6e0b6cafd01fc98a10
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/579280
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>