[codesize] Create diffs when run on the CI (post submit)
This also adds in a new perf datum - the diff between the previous
commit and this commit.

When we run the codesize taskdriver after a commit has landed,
we use the gitiles API to find the previous commit and use
that for the NoPatch build. Then, instead of ignoring the
NoPatch build, we use it to do a diff, in the same way we
do the CL diffing now.

Change-Id: I683f3dfcc4817c84246fa66fe4c36f3079b5236e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/581458
Reviewed-by: Eric Boren <borenet@google.com>
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>