[bazel] Resolve broken lib symlinks in Linux toolchain
Without this change, any code requiring the GPU (e.g.
hello_world_gl) would link without error but crash with
a confusing stacktrace, deep in the nvidia driver code

#0  0x00007ffff7c61c65 in realloc () from /lib/x86_64-linux-gnu/libc.so.6
#1  0x00007ffff796137f in ?? () from /lib/x86_64-linux-gnu/libxcb.so.1
#2  0x00007ffff7961404 in xcb_get_extension_data () from /lib/x86_64-linux-gnu/libxcb.so.1
#3  0x00007ffff76b0e8c in ?? () from /lib/x86_64-linux-gnu/libGLX_nvidia.so.0
...
#10 0x00007ffff7677927 in glXCreateContextAttribsARB () from /lib/x86_64-linux-gnu/libGLX_nvidia.so.0
#11 0x00007ffff7ae7ccc in ?? () from /lib/x86_64-linux-gnu/libGLX.so.0
#12 0x0000000000f615ff in (anonymous namespace)::GLWindowContext_xlib::onInitializeContext (this=0x12a9600)
    at tools/sk_app/unix/GLWindowContext_unix.cpp:91
...

As it turns out, debian packages like libx11-dev [1] have a
shared library like usr/lib/x86_64-linux-gnu/libX11.so, but this
is only a symlink to libX11.so.6.4.0, which is *not* included.
However, this package does include libX11.a, which the linker
will use after failing to find libx11.so.* [3][4].

Adding `--features=diagnostic_link` to the bazel build command
shows lld finding and using libX11.a, libxcb.a, and libXau.a
before this CL, and libX11.so, libxcb.so, libXau.so after.

`findelf -d [binary]` also shows a difference.
Before:
    Tag        Type                         Name/Value
    0x0000000000000001 (NEEDED)             Shared library: [libdl.so.2]
    0x0000000000000001 (NEEDED)             Shared library: [libGL.so.1]
    0x0000000000000001 (NEEDED)             Shared library: [libXdmcp.so.6]
    0x0000000000000001 (NEEDED)             Shared library: [libpthread.so.0]
    0x0000000000000001 (NEEDED)             Shared library: [libm.so.6]
    0x0000000000000001 (NEEDED)             Shared library: [libc.so.6]
    0x0000000000000001 (NEEDED)             Shared library: [ld-linux-x86-64.so.2]
    0x0000000000000001 (NEEDED)             Shared library: [librt.so.1]

After:

    Tag        Type                         Name/Value
    0x0000000000000001 (NEEDED)             Shared library: [libdl.so.2]
    0x0000000000000001 (NEEDED)             Shared library: [libGL.so.1]
    0x0000000000000001 (NEEDED)             Shared library: [libX11.so.6]
    0x0000000000000001 (NEEDED)             Shared library: [libxcb.so.1]
    0x0000000000000001 (NEEDED)             Shared library: [libXau.so.6]
    0x0000000000000001 (NEEDED)             Shared library: [libXdmcp.so.6]
    0x0000000000000001 (NEEDED)             Shared library: [libpthread.so.0]
    0x0000000000000001 (NEEDED)             Shared library: [libm.so.6]
    0x0000000000000001 (NEEDED)             Shared library: [libc.so.6]
    0x0000000000000001 (NEEDED)             Shared library: [ld-linux-x86-64.so.2]
    0x0000000000000001 (NEEDED)             Shared library: [librt.so.1]
    0x0000000000000015 (DEBUG)              0x0

Why did the linker succeed but runtime fail?
(speculation)
Probably one of the static libraries is not ABI compatible
with the dynamic version
(/speculation).

This CL resolves all broken symlinks in the hermetic toolchain
that I could find. //example:hello_world_gl, hello_world_vulkan,
and hello_world_dawn all work for me without crashing. The minimal
fix is probably just libxcb1 and possibly the other X11 libraries,
but I'd like to avoid this confusing issue in the future.

[1] https://packages.debian.org/bullseye/amd64/libx11-dev/filelist
[2] https://packages.debian.org/bullseye/amd64/libx11-6/filelist
[3] https://www.internalpointers.com/post/journey-across-static-dynamic-libraries
[4] https://www.ibm.com/docs/fr/cobol-aix/5.1?topic=files-linker-search-rules
Change-Id: Ia77b33b2922061d874a32f76ee54f82b6d1f3c8e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/582817
Reviewed-by: Ben Wagner <bungeman@google.com>