Use different procs for CPU/Ganesh/Graphite tests
While investigating how testing works as a part of Bazel-ifying
it, I realized 2 of the 3 test types ignore the GrContextOptions,
but this is not clear from the call sites.

This makes a different static constructor for each type
of test and a different TestProc function pointer type for
each as well. This allows test runners to more explicitly
handle tests of distinct types differently, and not
need filler or empty GrContextOptions.

Change-Id: Idc15eb5bda928410461b706635e46db0785eb261
Bug: skia:13758
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/583243
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>