Adjust gradient API in anticipation of color space options
We have long had a single flag (interpolate-in-premul). We're about to
add two enums to select the interpolation color space and hue method.
To manage this better, I'm adding an Interpolation struct, and
switching the internal storage on both the gradient shader and
descriptor. This will let us add more options without additional churn
on the API side.

Of course, clients are still using the flag API, so the old factories
are around, and everything now routes through the new factories.

Note that this still bakes the result down to the same serialized
format, so no SKP version tick is required (and even when we add the
color space options, we will do so in a way that makes legacy SKP
content behave the same, without a version check needed).

Bug: skia:13108
Change-Id: I2ff75d7beb4b091d2677b91a8007fafdfba02922
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/583958
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>