Remove redundant setReserve call
Having two of the same call increases the chance of having
two .eh_frame entries for the same functionality.

Notice that setReserve took an int, and reserve used to take
a size_t. Every site where reserve() was used was actually
passing an int. So, I just unified the calls to take an int.

Change-Id: I14ced1d51faf6dabff9b80ff1f6089db0c854e04
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/583967
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Herb Derby <herb@google.com>