[metal] Copy shader string into NSString to avoid invalid access.
It appears that when Metal caches shaders, it takes a ref to the
NSString passed in, rather than a copy. This means that when our
std::string goes out of scope they're referring an NSString with
deleted data. Copying it should fix this.

Bug: 1360743
Change-Id: Ia33bfd2ebfa273c3295a0693613eef4b5e5f19f2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/584456
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>