Store dst SkIRect and filter mode for copy surface tasks
This CL is the first in a series to make copySurface() support scaling
as part of the copy, which functions like glBlitFramebuffer can do.
When doing backdrop filters at a lower resolution, and the backdrop is
coming frome the framebuffer, scaling as part of the copy is more
efficient than copying the entire framebuffer and then downscaling.

This CL updates the backend-agnostic functionality of Ganesh to support
specifying a dst rect and filter mode to apply during the copy, and
updates the existing callers. Callers and caps checks don't take
advantage of differing dimensions yet, and the backend virtuals
will be updated. Once some backends support scaling copies,
SurfaceFillContext and SkDevice will be updated to take advantage.

Bug: b/197774543
Change-Id: Ieda65a9584715b72dc2c658d7b993c5dd927a7d5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/582397
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>