[graphite] Fix black areas in LCD text.
Colors in the LCD atlas end up getting sampled as (r,g,b,1), which
is not the correct alpha value for coverage. This adds the alpha
setup code for LCD that we use in Ganesh to handle this
situation.

Also adds the glyph conversion fixes from
https://skia-review.googlesource.com/c/skia/+/584919.

Bug: b/238757581
Change-Id: I5a8000721a19de869ade173b4cf1833a2ecfdf5e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/584921
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>