Invalidate CoreText font bounds for color glyphs
The conservative font bounds reported by CoreText are those in the font
and are meant primarially for (and are generally calculated only for)
the default instance of the glyph outlines in the font. In particular
they do not generally apply to `sbix` and `SVG ` data. When these tables
are present the conservative font bounds reported by the font should be
considered invalid since glyph representations used may not be outlines.

Bug: https://github.com/flutter/flutter/issues/108590
Change-Id: I9d591e4327839f0660413683bcc223d6b76849ef
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/585421
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Herb Derby <herb@google.com>