Include OptimizeForSize configuration in Perf trace keys
We are keeping the status quo for Release builds, as not to break the
history of all existing perf traces.

https://skia-review.googlesource.com/c/skia/+/585419 added
two CPU tests for our OptimizeForSize build (currently we don't
have the full GPU suite on this build, but we do have GPU
perf tests for skottie code specifically).

However, I couldn't find that data in Perf because it was labeled
the same as the normal release data.

Change-Id: Ice8d4fc3c3424d83106a8f046120feb832a6e997
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/585400
Reviewed-by: Ravi Mistry <rmistry@google.com>