Use unoptimized modules in debug builds for readability.
In a release build, or when optimizing for size, we use
`.minified.sksl` modules. These are packed as tightly as we can;
identifiers can be removed or renamed as long as it doesn't affect
the external behavior of the code.

Otherwise, we use `.unoptimized.sksl` modules. These modules are just
compiled and re-emitted without any meaningful changes, just basic
whitespace packing. (When the code is re-emitted into a shader, we'll
regenerate appropriate whitespace anyway.)

Change-Id: I8a8d507c6a3e41f48119f3d4a4dff7ecd2b87725
Bug: skia:13775
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/585741
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>