Fix simplified emboldened glyphs
Fake bold glyphs are done with a stroke+fill. The stroker creates a
winding stroke path and then appends the fill path. If the fill path was
even-odd this may be incorrect. Outlines in fonts are generally winding,
with a bit for requesting simplification. The Simplify method provided
by pathops generally returns an even-odd outline. For consistency and to
work with stroking, use AsWinding afer simplifying to produce the
simplified glyph path as a winding path.

Bug: https://github.com/flutter/flutter/issues/112546
Change-Id: Ic88ef10e5371b2f07c15f706d6aad1b8ab39f83c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/587877
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>