SkTDArray: for sizeof(T) == 1 force capacity to multiples of 16
SkTArrays frequently have a sizeof(T) of 1, causing the push_back
progression of capacity to be 7, 15, ... But, most systems of a
sizeof(max_align_t) == 16. So, there is no reason to allocate anything
less than 16 bytes. Have SkTDArrays round the capacity up to the next
multiple of 16.

Bug: b/249254511
Bug: chromium:1369069
Change-Id: I0ae36c6d1092484848bae6841844dfdfc8d96fde
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/588877
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>