Re-enable SkShaper tests
Since [0] the shaper tests have not been compiled or run since the
`SKSHAPER_IMPLEMENTATION` macro was only privately defined in the shaper
build itself and not defined for the test targets. This change moves the
test into the shaper module to better reflect how modules are now
structured and builds the shaper tests only when the shaper is being
built.

Re-enabling these tests discovered more issues, so the test is updated
to give more useful information in its asserts. These helped to debug an
issue in the shaper wrapping logic which would use the last glyph in a
cluster instead of the first to mark the cluster edges. This would cause
a wrapping opportunity before a multiple glyph cluster to incorrectly
include all but the last glyph in that cluster on the line.

[0] 8b2513a9866328b6cc393e3019ac58036e94d5ce
"Build the NDK compliant shared library to be used by JNI in SkQP."
https://skia-review.googlesource.com/c/skia/+/491447

Bug: skia:9050
Change-Id: I5b2b7dc2f7340496ecb5e69626c4686c6bc735e1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/589409
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>