Remove ParsedModule class.
ParsedModule was just a wrapper around a BuiltinMap and a SymbolTable.
It's reasonable to put the SymbolTable inside the BuiltinMap, and then
ParsedModule doesn't do anything at all.

ParsedModule contained two shared_ptrs and was passed around by value
sometimes; the new code uses unique_ptr instead, so it should be a
little lighter-weight.

(It's not unreasonable to take this logic a step further and ask,
why not merge LoadedModule and BuiltinMap into a single Module class?
The only significant distinction between a LoadedModule and a
BuiltinMap is that the LoadedModule preserves all of its
ProgramElements in the original order, and a BuiltinMap optimizes
for fast lookup by Symbol and discards FunctionPrototypes. I'll
think this over some more.)

Change-Id: I2a9e22aa07b071b3baa224561f41f0f4cfbef7e3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/589698
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>