Handle DirectWrite GetAlphaTextureBounds failure
It was previously known that at small sizes GetAlphaTextureBounds could
fail in an undocumented way and return invalid bounds.  It has been
discovered that this also happens at large sizes (but still within the
SkGlyph size range). When this happens the the DirectWrite
SkScalerContext would report that the glyph was of zero size when it was
actualy quite large.

Fix this by adding an additional fallback which uses Skia to generate
the metrics and image from the glyph path (if any).

Bug: chromium:1370488
Change-Id: Ied6009ff20a30f63503b004bd16c5299ffa5eb52
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/590181
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>