Harden SkCTFontGetDataFontWeightMapping
CoreText exposes a number of attibutes through `CFDictionary`, where all
the keys and values are `void*`. The expectation is that entries with
various keys will exist and have specific types, but nothing enforces
this so everything must be checked at runtime. In particular the
dictionary entry with key `kCTFontWeightTrait` is expected to have a
value of type `CFNumberRef` but when representing 0 this value has been
observed to sometimes have the type `CFBooleanRef` (with a `false`
value).

`SkCTFontGetDataFontWeightMapping` would previously return a default
mapping if this ever occurred, potentially leading to font weights being
mis-reported on fonts created from data. Instead, it should only fail if
the entry with key `kCTFontWeightTrait` does not exist. When the entry
with key `kCTFontWightTrait` exists consider the weight to be 0 if the
associated value is not a `CFNumberRef` or otherwise cannot be converted
to a `CGFloat`.

Since this change could lead to issues where all weights are seen as 0,
also enforce that the weights must be strictly monotonically increasing
to be valid. This should have been checked previously as all users of
the returned array assume this property.

Bug: chromium:1372420
Change-Id: I4ad1d7066bad69331064e135fb015f44ffc37a2b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/590553
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>