Look up built-in global variables without BuiltinMap.
Previously, we had a mechanism to go from a Variable to its
VarDeclaration statement. Howver, there was no mechanism to go from
the VarDeclaration (statement) to the associated GlobalVarDeclaration
(program element). This presented a problem, since we needed to look
up GlobalVarDeclarations from a Variable to copy the built-in data.

This CL updates the declaration pointer in Variable so that it points
at a GlobalVarDeclaration for global variables, or a VarDeclaration
for non-global variables. An IRNode* can reference either, and we can
differentiate between the two by checking the kind. We now have
accessors to return the VarDeclaration* or GlobalVarDeclaration*
by checking the IRNode's kind; this has a very low runtime cost (one
comparison) and no additional storage cost.

Now that we have a mechanism for going from a Variable to the
associated GlobalVarDeclaration, we can use it to set up built-in
variables without involving the BuiltinMap. This doesn't address
interface blocks and fields, though; that will be done in a
separate CL (this is already large enough).

Change-Id: I5e2cbb60ce4c41362ef45f9e57055e70b9b658c6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/590736
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>