Return a pointer from VarDeclaration::var().
Originally, variable declarations were designed so that the variable
pointer inside should never be null. However, there is a brief window
where this is not true--when a Variable is being deleted (e.g.
it's optimized away or we are destroying the Program), it tells the
VarDeclaration to clear its pointer, to avoid any stale pointer
accesses.

Changing the `var()` accessor to return a pointer will allow me to
detect this case. instead of accessing a null reference (in a followup
CL).

Change-Id: Icaba3057d5755a7a073c3834e8d89271147ebdb7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/591000
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>