Remove triangulating path renderer for SK_ENABLE_OPTIMIZE_SIZE
Stripped skottie_tool 17095256 -> 17042008 = -52kb.

This removes GrTriangulator and its subclasses GrAATriangulator and
GrInnerFanTriangulator. It removes the TriangulatingPathRenderer from
the path chain, which relied on GrTriangulator (for MSAA) and
GrAATriangulator (for analytic AA).

This also removes PathInnerTriangulateOp and modifies the logic for
selecting an op in TessellationPathRenderer to skip the variant that
would use GrInnerFanTriangulator for its inner triangles.

Tests, benchmarks, samples, and fuzzes also had to be tweaked to no
longer reference the triangulator types.

Bug: skia:13657
Change-Id: If54b7a40ce03c5487a1f4e6ed994f82f2385f99f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/591077
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>