Limit multi-sampling when using implicit resolve
This is a case where Angle validation caught a misuse by skia.

The error is triggered when trying to allocate a render buffer storage with greater number of samples than the format allows. We determine the max multisample count in skia [2] via per format caps. The incorrect usage case is that multisampling of implicit buffers uses a different set of caps  than the per format multisample [3].


[1]
https://source.chromium.org/chromium/chromium/src/+/main:third_party/angle/src/libANGLE/validationES2.cpp;drc=b44e2b7a8a0df47f5a2d114d82ff8d5375e13bd7;bpv=1;bpt=1;l=6469?q=%22Samples%20must%20not%20be%20greater%20than%20maximum%20supported%20value%20for%20the%20format%22

[2]
https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/src/gpu/ganesh/GrDynamicAtlas.cpp;drc=b44e2b7a8a0df47f5a2d114d82ff8d5375e13bd7;bpv=1;bpt=1;l=65?q=GrCaps.h&ss=chromium%2Fchromium%2Fsrc

[3]
https://developer.download.nvidia.com/opengl/specs/GL_EXT_framebuffer_multisample.txt

Bug: 1353374
Change-Id: Idc686c5abe55f1dc4722eca2238a2dd9f8893878
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/590540
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>