Guard against very small and very large font sizes
Change the guard bypass SDFT and DirectMask when the font is very
large. Skip the drawing of last resort if the size of the font is
too small. If the font is very small, it could return 0 for
maxDimension causing an infinite loop in the matrix conditioning code.

Bug: chromium:1370488
Change-Id: I0531a5ec40955161828cb17399bf40a6877c5ffa
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/591267
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Herb Derby <herb@google.com>