Converts SkPathRef to use SkSTArray
I settled on the values by running MotionMark with a bunch of
permutations. The suites that are impacted by this are arcs, paths,
and lines. If I look purely at lines, I would go with 2,2,2. But
going with 4,4,2 seems to give the most benefit to all.

I had to tweak SkTArray::operator=(SkTArray&&). Without this, the
test pathbuilder_shrinkToFit fails. The reason the test would fail is
that the test is assuming the size returned from snapshot() to be
less than that of detach(). The problem is, the constructor for
SkStArray would end up allocating more, and the test failing. The
alternative is to nuke the test.

Bug: skia:13817


Change-Id: Id56858b2362bc5288a03e3ad9d4d96692133c042
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/590698
Reviewed-by: Herb Derby <herb@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Herb Derby <herb@google.com>