Disallow private identifiers completely in user code.
The parser now has a dedicated token type for private identifiers.
In a public runtime effect ProgramKind, finding a private identifier
is now an immediate parse error. In other ProgramKinds, they're
always allowed. This seems like what we want in practice, considering
the way we are starting to use $-prefixes on all sorts of built-in
functions.

This change allowed some existing $-prefix error checks to be removed.
The error messages on the existing checks weren't entirely consistent,
so this change had a slight ripple effect on some of the test output.

Change-Id: I4463de00228150111f97d9ba5884ce8ae8b8028d
Bug: skia:13810
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/591270
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>